fixed b[]

+++ make_wav_gen1.c
@@ -38,8 +38,8 @@ static char decode(int word, int flip);
 static void printAll10bwords();
 static int encode(int word);
 static int twosToInt(int val,int len);
-static int encodeB(uint8_t *b, int index, int val);
-static int encodeA(uint8_t *b, int index, int val);
+static int encodeB(short int  *b, int index, int val);
+static int encodeA(short int  *b, int index, int val);
 
 	 static  int NOT_FRAME = /* 0fa */ 0xfa & 0x3ff;
 	 static  int FRAME = /* 0fa */ ~0xfa & 0x3ff;
@@ -1271,14 +1271,14 @@ void write_to_buffer(int i, int symbol, int val)
 		return -99; // error state - we did not find the RD sense
 	}
 	
-int encodeA(uint8_t *b, int index, int val) {
+int encodeA(short int  *b, int index, int val) {
 //    printf("Encoding A\n");
     b[index] = (val & 0xff);
     b[index + 1] = b[index + 1] | (val >> 8);
     return 0;	
 }
 
-int encodeB(uint8_t *b, int index, int val) {
+int encodeB(short int  *b, int index, int val) {
 //    printf("Encoding B\n");
     b[index] = b[index]  |  (val << 4);
     b[index + 1] = b[index + 1]  |  ((val >> 4 ) & 0xff);
